Ensure basic memory types work on all architectures
* Clean up `memory_structs` to remove unnecessary content
  specific to multiboot2 or the x86_64 architecture.
