+++ kernel/memory_structs/Cargo.toml
@@ -1,12 +1,10 @@
 [package]
 authors = ["Kevin Boos <kevinaboos@gmail.com>"]
 name = "memory_structs"
-description = "Common types used in the memory management subsystem"
+description = "Basic types used for memory management"
 version = "0.1.0"
 
 [dependencies]
-multiboot2 = "0.14.0"
-xmas-elf = { version = "0.6.2", git = "https://github.com/theseus-os/xmas-elf.git" }
 bit_field = "0.7.0"
 zerocopy = "0.5.0"
 derive_more = "0.99.0"
@@ -14,9 +12,6 @@ paste = "1.0.5"
 
 [dependencies.kernel_config]
 path = "../kernel_config"
-
-[dependencies.entryflags_x86_64]
-path = "../entryflags_x86_64"
 
 [lib]
 crate-type = ["rlib"]
